emails/requirements.txt
fastapi
uvicorn
python-dotenv
emails
bcrypt